Firebase Database integration done. Favourite Rockets page design, dev done.

+++ app/src/main/java/com/example/spacexfan/model/modelproperties/Mass.kt
@@ -6,6 +6,6 @@ import java.io.Serializable
 
 data class Mass (
 
-   @SerializedName("kg") var kg : Int
+   @SerializedName("kg") var kg : Long
 
 ) : Serializable
